+++ app/src/main/java/com/example/navsample/entities/relations/PriceByCategory.kt
@@ -1,7 +1,7 @@
 package com.example.navsample.entities.relations
 
 data class PriceByCategory(
-    var price: Double,
+    var price: Int,
     var category: String,
     var date: String,
 )
